+++ gradle.properties
@@ -1,10 +1,10 @@
 # IntelliJ Platform Artifacts Repositories
 # -> https://plugins.jetbrains.com/docs/intellij/intellij-artifacts.html
 
-pluginGroup = com.github.nggalien.ideaplpgdebugger
+pluginGroup = net.plpgsql.ideadebugger
 pluginName = idea-plpgdebugger
 # SemVer format -> https://semver.org
-pluginVersion = 0.0.1
+pluginVersion = 1.0.0-alpha
 
 # See https://plugins.jetbrains.com/docs/intellij/build-number-ranges.html
 # for insight into build numbers and IntelliJ Platform versions.
