+++ src/main/kotlin/net/plpgsql/ideadebugger/Logger.kt
@@ -2,7 +2,7 @@
  * Copyright (c) 2022. Alexandre Boyer
  */
 
-package net.plpgsql.debugger
+package net.plpgsql.ideadebugger
 import com.intellij.openapi.diagnostic.Logger
 import kotlin.reflect.KClass
