+++ src/main/kotlin/net/plpgsql/ideadebugger/PlProcess.kt
@@ -2,9 +2,9 @@
  * Copyright (c) 2022. Alexandre Boyer
  */
 
-package net.plpgsql.debugger
+package net.plpgsql.ideadebugger
 
-import com.intellij.database.dataSource.DatabaseConnectionPoint
+import com.intellij.database.dataSource.DatabaseConnection
 import com.intellij.database.debugger.SqlDebugProcess
 import com.intellij.database.debugger.SqlDebuggerEditorsProvider
 import com.intellij.openapi.project.Project
@@ -17,64 +17,83 @@ import com.intellij.xdebugger.breakpoints.XLineBreakpoint
 import com.intellij.xdebugger.evaluation.XDebuggerEditorsProvider
 import com.intellij.xdebugger.frame.XExecutionStack
 import com.intellij.xdebugger.frame.XSuspendContext
-import kotlin.properties.Delegates
+import org.jetbrains.concurrency.runAsync
 
 
 class PlProcess(
     session: XDebugSession,
-    private val connectionPoint: DatabaseConnectionPoint,
+    val connection: DatabaseConnection,
     private val entryPoint: Long
 ) : SqlDebugProcess(session) {
 
     private val logger = getLogger<PlProcess>()
     private val editor = EditorProvider()
     private val context = XContext()
-    val connection = getConnection(session.project, connectionPoint)
     var sessionId: Int = 0
-
     private val stack = XStack(session)
+    var step: PlStep? = null
+    var aborted: Boolean = false
 
     override fun startStepOver(context: XSuspendContext?) {
         logger.debug("startStepOver")
-        plStepOver(connection, sessionId)
-        reached()
+        goToStep(Request.STEP_OVER)
+
     }
 
     override fun startStepInto(context: XSuspendContext?) {
         logger.debug("startStepInto")
-        plStepInto(connection, sessionId)
-        reached()
+        goToStep(Request.STEP_INTO)
     }
 
     override fun resume(context: XSuspendContext?) {
         logger.debug("resume")
-        plContinue(connection, sessionId)
+        goToStep(Request.STEP_CONTINUE, 500)
     }
 
     override fun sessionInitialized() {
         super.sessionInitialized()
         if (entryPoint == 0L) {
-            connection.remoteConnection.close()
-            stop()
+            session.stop()
         }
     }
 
     fun startDebug(port: Int) {
         sessionId = plAttach(connection, port) ?: 0
-        reached()
+        stack.update(step)
+        session.positionReached(context)
     }
 
     override fun stop() {
         super.stop()
-        connection.remoteConnection.close()
+        try {
+            if (!connection.remoteConnection.isClosed && !aborted) {
+                abort().blockingGet(100)
+            }
+        } catch (e: Exception) {
+            logger.debug(e)
+        } finally {
+            connection.remoteConnection.close()
+        }
     }
 
-    private fun reached() {
-        stack.update()
-        session.positionReached(context)
+    private fun goToStep(request: Request, timeOut: Int = 300) {
+        try {
+            step = fetchStep(request).blockingGet(timeOut)
+            stack.update(step)
+            session.positionReached(context)
+        } catch (e: Exception) {
+            aborted = true
+            session.stop()
+        }
     }
 
-    fun getStack() = plGetStack(connection, sessionId)
+    private fun fetchStep(request: Request) = runAsync {
+        plRunStep(sessionId, connection, request)
+    }
+
+    private fun abort() = runAsync {
+        plAbort(connection, sessionId)
+    }
 
     override fun getEditorsProvider(): XDebuggerEditorsProvider {
         return editor
@@ -117,6 +136,7 @@ class PlProcess(
     }
 
     inner class XContext : XSuspendContext() {
+
         override fun getActiveExecutionStack(): XExecutionStack? {
             return stack
         }
